# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ipoh Drum Laravel API",
        "description": "This is a swagger-generated API documentation for the project Ipoh Drum. (Only supports OpenAPI Annotations for now.)",
        "contact": {
            "email": "henry_lcz97@hotmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "The URL that IpohDrum Laravel API is running on."
        }
    ],
    "paths": {
...
        "/api/companytype": {
            "get": {
                "tags": [
                    "CompanyTypeControllerService"
                ],
                "summary": "Get list of companytypes",
                "description": "Returns list of companytypes",
                "operationId": "getCompanyTypes",
                "parameters": [
                    {
                        "name": "pageNumber",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "number of pageSize",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved list of companytypes"
                    },
                    "default": {
                        "description": "Unable to retrieve list of companytypes"
                    }
                }
            },
            "post": {
                "tags": [
                    "CompanyTypeControllerService"
                ],
                "summary": "Creates a companytype.",
                "operationId": "createCompanyType",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "CompanyType Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "desc",
                        "in": "query",
                        "description": "CompanyType Description",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CompanyType has been created successfully."
                    },
                    "default": {
                        "description": "Unable to create the companytype."
                    }
                }
            }
        },
        "/api/filter/companytype": {
            "get": {
                "tags": [
                    "CompanyTypeControllerService"
                ],
                "summary": "Filter list of companytypes",
                "description": "Returns list of filtered companytypes",
                "operationId": "filterCompanyTypes",
                "parameters": [
                    {
                        "name": "pageNumber",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "number of pageSize",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Keyword for filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fromdate",
                        "in": "query",
                        "description": "From Date for filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "todate",
                        "in": "query",
                        "description": "To string for filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "status for filter",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved list of filtered companytypes"
                    },
                    "default": {
                        "description": "Unable to retrieve list of companytypes"
                    }
                }
            }
        },
        "/api/companytype/{uid}": {
            "get": {
                "tags": [
                    "CompanyTypeControllerService"
                ],
                "summary": "Retrieves companytype by Uid.",
                "operationId": "getCompanyTypeByUid",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "CompanyType_ID, NOT 'ID'.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CompanyType has been retrieved successfully."
                    },
                    "default": {
                        "description": "Unable to retrieve the companytype."
                    }
                }
            },
            "put": {
                "tags": [
                    "CompanyTypeControllerService"
                ],
                "summary": "Update companytype by Uid.",
                "operationId": "updateCompanyTypeByUid",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "CompanyType_ID, NOT 'ID'.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "CompanyTypename",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "desc",
                        "in": "query",
                        "description": "CompanyType Description",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CompanyType has been updated successfully."
                    },
                    "default": {
                        "description": "Unable to update the companytype."
                    }
                }
            },
            "delete": {
                "tags": [
                    "CompanyTypeControllerService"
                ],
                "summary": "Set companytype's 'status' to 0.",
                "operationId": "deleteCompanyTypeByUid",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "CompanyType ID, NOT 'ID'.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CompanyType has been 'deleted' successfully."
                    },
                    "default": {
                        "description": "Unable to 'delete' the companytype."
                    }
                }
            }
        },
...
    },
...
}